--- src/main/java/student_oleg_gryazev/_lesson_3/DemoTrafficLightControlsTheCar.java
@@ -1,47 +1,0 @@
-package student_oleg_gryazev._lesson_3;
-
-public class DemoTrafficLightControlsTheCar {
-    public static void main(String[] args) {
-        TrafficLightControlsTheCar car = new TrafficLightControlsTheCar();
-        TrafficLightControlsTheCar trafficLight = new TrafficLightControlsTheCar();
-
-        car.carPosition();
-        trafficLight.positionOfTrafficLight();
-    }
-
-}
-class Computer {
-
-    String manufacturer;
-    String model;
-
-    Computer(String manufacturer, String model) {
-        this.manufacturer = manufacturer;
-        this.model = model;
-
-    }
-    public String getManufacturer(){
-      return manufacturer;
-    }
-    public String getModel(){
-        return model;
-    }
-
-
-
-public void setManufacturer(String manufacturer){
-
-}
-}
-
-class ComputerDemo {
-
-    public static void main(String[] args) {
-        Computer dell = new Computer("Dell", "XXX-1");
-        String manufacturer = dell.getManufacturer();
-        String model = dell.getModel();
-        System.out.println("Computer manufacturer = " + manufacturer);
-        System.out.println("Computer model = " + model);
-    }
-
-}
