--- src/main/java/student_oleg_gryazev/_lesson_3/level_7/cirkleDemo.java
@@ -1,9 +1,0 @@
-package student_oleg_gryazev._lesson_3.level_7;
-
- class CirkleDemo {
-    public static void main(String[] args) {
-        Circle circle = new Circle();
-        circle.calculateArea(5);
-
-    }
-}
