+++ src/main/java/student_oleg_gryazev/_lesson_3/level_7/task_30/cirkleDemo.java
@@ -1,0 +1,11 @@
+package student_oleg_gryazev._lesson_3.level_7.task_30;
+
+import student_oleg_gryazev._lesson_3.level_7.task_30.Circle;
+
+class CirkleDemo {
+    public static void main(String[] args) {
+        Circle circle = new Circle();
+        circle.calculateArea(5);
+
+    }
+}
