+++ src/main/java/student_oleg_gryazev/_lesson_3/level_7/task_33/Computer.java
@@ -1,0 +1,28 @@
+package student_oleg_gryazev._lesson_3.level_7.task_33;
+
+ class Computer {
+
+        String manufacturer;
+        String model;
+
+        Computer(String manufacturer, String model) {
+        this.manufacturer = manufacturer;
+        this.model = model;
+
+    }
+
+        public String getManufacturer() {
+        return manufacturer;
+    }
+
+        public String getModel() {
+        return model;
+    }
+
+
+        public void setManufacturer(String manufacturer) {
+
+    }
+    }
+
+
