+++ src/main/java/student_oleg_gryazev/_lesson_3/level_7/task_33/ComputerDemo.java
@@ -1,0 +1,6 @@
+package student_oleg_gryazev._lesson_3.level_7.task_33;
+
+public class ComputerDemo {
+
+    }
+}
